+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For a more detailed list of past and incoming changes, see the commit history.
 - The strength of sculpting tools is now proportional to brush size
 - Terrain resources generate a greyscale thumbnail
 - Fixed potential cleanup error when the editor closes
+- Fixed custom types not cleaned up when the plugin is disabled
 - Fixed LOD to take height into account on tall mountains